Update our auto sell to work with 1.2
Instead of a delay, items are sold as soon as you loot them (if enabled). This circumvents the need for updating the inventory imagery (saving quite a bit of processing). Later on the script should be configured to NOT sell items you had equipped for a certain
amount of time, possibly a minute or two (or never?). That's to be decided.